+++ tests/posix/fs/src/test_fs_dir.c
@@ -6,6 +6,7 @@
 
 #include <stdio.h>
 #include <posix/unistd.h>
+#include <posix/dirent.h>
 #include "test_fs.h"
 
 extern int test_file_write(void);
